# proto: fix quadratic behavior in nested map marshaling